Fix the interrupts in EL3
The reason that the interrupts in EL3 were not being handled it is because
the code was linked at address 0x80000 instead of 0x0. When an interrupt was
triggered, the interrupt handlers were not found in the expected location.

Because we are starting in EL3, the attribute "kernel_old" is set to 1. This
causes that the kernel be loaded at 0x0 instead of 0x80000.

See the link below for more information:

https://www.raspberrypi.org/documentation/configuration/config-txt/boot.md